WIP Fix bugs with half carry and overflow
* Fixes an issue in ADD HL where the flags weren't evaluated with the
upper 16 bits.

There are a lot of minor changes here but in an attempt to avoid minor
differences between instructions helper functions for half carry
addition and subtraction are used.

There is still a lot of refactoring required in porting the codebase
from the 8080 (eighty-eighty), such as the DAA instruction needs to
support subtraction as well.

Still not passing CPUTEST, but most common instructions should be
functional.