Fix issue with ADD HL and flag check upper byte
This hopefully fixes an issue with the HF flag not being set on carries
occuring from `FFh` to `00h` and HL, from `00FFh` to `0100h`.
Added tests to hopefully prevent further regression in this area.

Passes the preliminary z80 tests.

A LOT of refactoring is necessary here to slim down the codebase.