Now passes the CPUTEST and preliminary z80 tests
Refactored & removed unused functions.
Pacman subdirectory for display, keypad input etc..
TODO: Move z80 code into separate crate.

* Started work on memory mapping for Pacman
Interrupts still need to be handled.

Basically if in VBLANK or if bit 0 of 5000h is 1 an interrupt can happen
and this needs to be signalled to the CPU, irq enabled etc..