Implement CPM flag to toggle memory mapping
We need to handle this better but from the original ported code
everything was stored in one vec, which means we were reading and
writing to the same area. Not good.

In this case it's fine for the CPM tests but you want to read from ROM
and write to RAM, which was implemented with regards to Pacman but not
tests.. For now a simple flag will suffice but it's not ideal.

Other changes:

* Minor work done on the graphics end.
* Cleaned up the load method where we load the rom files.
* Shortened RST function.*

RST*: We had a bug here due to using the `write16` function it would
attempt to write to addresses in ROM space, however instead we just
index rom here just like `CALL` would for the vector.