More extended instructions & fix possible bug in LD HL, r)
Opcode 0x77:

```
self.memory[addr] = self.read_reg(src);
self.write_reg(src, self.memory[self.get_pair(HL)]);
```
We were writing the value of the register to the memory location of HL,
then writing the value of HL indexed in memory to the src register...?
That's not right.

TODO: Might want to clean up Index / IndexMut impl because we're
technically writing to ROM, although this isn't an issue for tests or
CP/M it's still nicer to make this more explicit to avoid bugs.